third_party/WebKit/LayoutTests/platform/win/editing/selection/select_all/select_all_input-expected.txt

This is a testharness.js-based test.
PASS SelectAll is not enabled if input is empty 
Harness: the test ran to completion.
